Assaignment/ass-1/pyproject.toml
[project]
name = "new-folder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.0.18",
]
